shared/shared/requirements.txt
pydantic==2.12.5
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
pymongo==4.9.1
motor==3.6.0
aio-pika==9.4.1
python-json-logger==2.0.7
redis==5.2.1

